[Oguzcan BICER]-add Global Exception Handler

+++ src/main/java/com/oguzcan/securitydemo/service/CourseService.java
@@ -1,13 +1,14 @@
 package com.oguzcan.securitydemo.service;
 
 import com.oguzcan.securitydemo.dto.request.CreateCourseRequestDTO;
+import com.oguzcan.securitydemo.exception.NotFoundException;
 import com.oguzcan.securitydemo.mapper.CourseMapper;
 import com.oguzcan.securitydemo.model.Course;
 import com.oguzcan.securitydemo.repository.CourseRepository;
-import com.oguzcan.securitydemo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -18,7 +19,8 @@ public class CourseService {
 
     private final CourseRepository courseRepository;
     private final CourseMapper courseMapper;
-    public Course createCourse (CreateCourseRequestDTO request) {
+
+    public Course createCourse(CreateCourseRequestDTO request) {
         Course course = courseMapper.toCourse(request);
         return courseRepository.save(course);
     }
@@ -29,7 +31,7 @@ public class CourseService {
 
     public Course getCourse(long courseId) {
         return courseRepository.findById(courseId)
-                .orElseThrow();
+                .orElseThrow(() -> new NotFoundException("Id:" + courseId + " -> Kurs bulunamadı."));
     }
 
     public Course setPassiveCourse(long courseId) {
@@ -42,7 +44,11 @@ public class CourseService {
         return courseRepository.save(updatedCourse);
     }
 
+    @Transactional
     public void deleteCourse(long courseId) {
+        var isNotExist = !courseRepository.existsById(courseId);
+        if (isNotExist)
+            throw new NotFoundException("Id:" + courseId + " -> Silinecek kurs bulunamadı.");
         courseRepository.deleteById(courseId);
     }
 }
